[WIP] Cache config
