+++ pom.xml
@@ -135,6 +135,20 @@
 			<artifactId>hypersistence-utils-hibernate-63</artifactId>
 			<version>3.10.1</version>
 		</dependency>
+
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-cache</artifactId>
+			<version>3.1.5</version>
+		</dependency>
+
+		<dependency>
+			<groupId>com.github.ben-manes.caffeine</groupId>
+			<artifactId>caffeine</artifactId>
+			<version>3.2.2</version>
+		</dependency>
+
+
 	</dependencies>
 
 	<build>
